refactor(cli): extract stop/upgrade/migration into separate lib modules
- node.ts: 368→88 lines, now pure command routing
- node-start.ts: inline migration guard replaced with shared runMigrationGuard()
- New lib/node-migration.ts: shared migration guard for start and upgrade
- New lib/node-stop.ts: extracted stop command implementation
- New lib/node-upgrade.ts: extracted upgrade command implementation
